better screenshot + merge cod
